2006-08-17 [colin]	2.4.0cvs57
	* src/imap.c
		Add View Log button on auth error
	* src/procmsg.c
		Fix encrypted mail marked as new when
		saved in normal folders
	* src/plugins/pgpinline/pgpinline.c
		Fix encrypt bug on IMAP
	* src/mimeview.c
	* src/common/utils.c
	* src/common/utils.h
	* src/gtk/inputdialog.c
	* src/gtk/inputdialog.h
		Get rid of metamail stuff, simplify
		opening, and allow remembering of user's
		choices (by updating ~/.mailcap). There's
		still to drop "Audio player" and "Image
		viewer" preferences.

+++ src/plugins/pgpinline/pgpinline.c
@@ -585,7 +585,8 @@ static gboolean pgpinline_sign(MimeInfo *mimeinfo, PrefsAccount *account)
 
 	if (msgcontent->content == MIMECONTENT_FILE &&
 	    msgcontent->data.filename != NULL) {
-		g_unlink(msgcontent->data.filename);
+	    	if (msgcontent->tmp == TRUE)
+			g_unlink(msgcontent->data.filename);
 		g_free(msgcontent->data.filename);
 	}
 	msgcontent->data.mem = g_strdup(tmp);
@@ -694,7 +695,8 @@ static gboolean pgpinline_encrypt(MimeInfo *mimeinfo, const gchar *encrypt_data)
 
 	if (msgcontent->content == MIMECONTENT_FILE &&
 	    msgcontent->data.filename != NULL) {
-		g_unlink(msgcontent->data.filename);
+	    	if (msgcontent->tmp == TRUE)
+			g_unlink(msgcontent->data.filename);
 		g_free(msgcontent->data.filename);
 	}
 	msgcontent->data.mem = g_strdup(tmp);
